Rename XTime > XDate and move types into their own go files
