+++ excellent/types/base_test.go
@@ -1,10 +1,8 @@
 package types_test
 
 import (
-	"encoding/json"
 	"fmt"
 	"testing"
-	"time"
 
 	"github.com/nyaruka/goflow/excellent/types"
 
@@ -21,7 +19,7 @@ func TestCompare(t *testing.T) {
 		{nil, nil, 0, false},
 		{nil, types.NewXString(""), 0, true},
 		{types.NewXError(fmt.Errorf("Error")), types.NewXError(fmt.Errorf("Error")), 0, false},
-		{types.NewXError(fmt.Errorf("Error")), types.XTimeZero, 0, true}, // type mismatch
+		{types.NewXError(fmt.Errorf("Error")), types.XDateZero, 0, true}, // type mismatch
 		{types.NewXString("bob"), types.NewXString("bob"), 0, false},
 		{types.NewXString("bob"), types.NewXString("cat"), -1, false},
 		{types.NewXString("bob"), types.NewXString("ann"), 1, false},
@@ -40,34 +38,4 @@ func TestCompare(t *testing.T) {
 			assert.Equal(t, test.result, result, "result mismatch for inputs '%s' and '%s'", test.x1, test.x2)
 		}
 	}
-}
-
-func TestXNumberMarshaling(t *testing.T) {
-	// unmarshal with quotes
-	var num types.XNumber
-	err := json.Unmarshal([]byte(`"23.45"`), &num)
-	assert.NoError(t, err)
-	assert.Equal(t, types.RequireXNumberFromString("23.45"), num)
-
-	// unmarshal without quotes
-	err = json.Unmarshal([]byte(`34.56`), &num)
-	assert.NoError(t, err)
-	assert.Equal(t, types.RequireXNumberFromString("34.56"), num)
-
-	// marshal (doesn't use quotes)
-	data, err := json.Marshal(types.RequireXNumberFromString("23.45"))
-	assert.NoError(t, err)
-	assert.Equal(t, []byte(`23.45`), data)
-}
-
-func TestXTimeMarshaling(t *testing.T) {
-	var date types.XTime
-	err := json.Unmarshal([]byte(`"2018-04-09T17:01:30Z"`), &date)
-	assert.NoError(t, err)
-	assert.Equal(t, types.NewXTime(time.Date(2018, 4, 9, 17, 1, 30, 0, time.UTC)), date)
-
-	// marshal
-	data, err := json.Marshal(types.NewXTime(time.Date(2018, 4, 9, 17, 1, 30, 0, time.UTC)))
-	assert.NoError(t, err)
-	assert.Equal(t, []byte(`"2018-04-09T17:01:30Z"`), data)
 }
